Verified Serial Functionality
Code includes measures on the transmitter end and receiver end to test
the functionality of this serial communication.  Namely a button on the
TX end and an LED on the RX end.  The communication is tested for
switching between two values in real time.  Current code has potential
to send 4 integers over.