Clean Working Encoder Code
This ought to be the encoder code the way we envisioned it using the
motorcontrol.cpp and just feeding it a value.  It uses only one encoder
counting just 16 revolutions in the circle.  It appears to stop the code
at close to the desired amount, plus a constant added bit that results
from the momentum of the motor